вынесен render.js с replace и renderDom
